move useSelector in the component
