--- src/hooks/index.js
@@ -1,4 +1,0 @@
-export { useToken } from '@plone/volto/hooks/userSession/useToken';
-export { useActions } from '@plone/volto/hooks/actions/useActions';
-export { useTypes } from '@plone/volto/hooks/types/useTypes';
-export { useContent } from '@plone/volto/hooks/content/useContent';
